refactor(dashboard): trim gateway compatibility wrappers
Why this change was needed:
Follow-up review found a few compatibility wrappers left behind after the gateway sync and release boundary refactor. Keeping them made the new ownership less clear and preserved imports through the old gateway facade.

What changed:
- Removed the unused module-level sync_gateway wrapper and export
- Removed release symbols from the biz.gateway re-export surface
- Moved sync_related_apps onto GatewayRelatedAppHandler and updated callers/tests
- Pointed release-related imports at biz.release.gateway_releaser directly

Problem solved:
The refactor now has fewer dead exports, clearer handler ownership for related-app sync, and release behavior no longer leaks through the gateway package.
